// src/jobs/types/job-action.ts
export enum JobAction {
  NONE = 'NONE',
  RESGITER = 'RESGITER',
  ENTER_PW_FOR_REGISTER = 'ENTER_PW_FOR_REGISTER',
  ENTER_PW_FOR_CHANGE_PW = 'ENTER_PW_FOR_CHANGE_PW',
  ENTER_PK_FOR_IMPORT_W = 'ENTER_PK_FOR_IMPORT_W',
  ENTER_PW_FOR_IMPORT_W = 'ENTER_PW_FOR_IMPORT_W',
  ENTER_PK_FOR_REPLACE_W = 'ENTER_PK_FOR_REPLACE_W',
  ENTER_PW_FOR_REPLACE_W = 'ENTER_PW_FOR_REPLACE_W',
  ENTER_PW_FOR_SHOW_PK = 'ENTER_PW_FOR_SHOW_PK',
  ENTER_DEPOSIT_AMOUNT = 'ENTER_DEPOSIT_AMOUNT',
  ENTER_PW_FOR_DEPOSIT = 'ENTER_PW_FOR_DEPOSIT',
  ENTER_WITHDRAW_SECRET_KEY = 'ENTER_WITHDRAW_SECRET_KEY',
  ENTER_WITHDRAW_RECIPIENT = 'ENTER_WITHDRAW_RECIPIENT',
  ENTER_PW_FOR_WITHDRAW = 'ENTER_PW_FOR_WITHDRAW',
  ENTER_SWAP_TOKEN = 'ENTER_SWAP_TOKEN',
  ENTER_SWAP_SECRET_KEY = 'ENTER_SWAP_SECRET_KEY',
  ENTER_SWAP_RECIPIENT = 'ENTER_SWAP_RECIPIENT',
  ENTER_SWAP_SLIPPAGE = 'ENTER_SWAP_SLIPPAGE',
  ENTER_PW_FOR_SWAP = 'ENTER_PW_FOR_SWAP',
}